+++ CHANGELOG.md
@@ -6,3 +6,17 @@
 - Default template.
 - HTTP generation endpoint.
 - Example client.
+
+## [1.1.0] - 2026-05-10
+
+### Added
+
+- Email capability with email generation endpoint.
+- Introduce gradle capabilities for modular dependencies.
+
+## [1.2.0] - 2026-05-14
+
+### Added
+
+- Logging of configuration at startup.
+- CLI capability with command line generation.